builtin/k8s: Do not preserve old annotations
The previous fix for ingress annotations would accidentally preserve any
old annotations that might of been set in a waypoint.hcl. So if a user
decided to update and remove some old values, they would continue to be
present in the next release. This commit fixes that by setting the
ingress resource annotations to whats exactly set in a waypoint.hcl.