internal/ceb/ssh: wait on exit until stdout copy is complete
There is a race condition between the command exiting and the copy to
stdout being complete. On command completion, we want to wait for the
stdout copy to finish before actually exiting.

This also fixes our #1 flaky test, which I believe was not in fact false
positive flaky, but was REAL positive flaky; I think this was a bug.