Increase failure threshold from 5 to 20.
Currently, by default, waypoint is timing out the deployment if the pod isn't ready in under a minute. This often isn't enough time, especially for an initial deployment that needs to download a docker image, or for large spring-boot apps that take a while to initialize.

Increasing the failure threshold to 30 will increase the timeout to 5 minutes, which should be enough time the majority of apps to pull and start. Users with quick-starting apps (or _very_ slow apps) can change this in their k8s config by modifying pod.container.failure_threshold: https://tip.waypointproject.io/plugins/kubernetes#pod-container-probe-failure_threshold

This is a tough balance - it isn't uncommon to deploy an app that won't ever become healthy, and users often won't start investigating until waypoint times out and reports the failure. If we time out sooner, users can start debugging sooner, and they would appreciate that cycle being faster.

I think that waypoint giving up too early on a working app is a pretty bad failure mode though - it isn't obvious what the problem is, or how to fix itby changing config.