Always set a project ref if we're loading config.
We know we'll always have one there, and it's required in some unexpected places, like the `config get` command. As far as I've found, there isn't a place where we look for the project ref being nil to do some project-agnostic behavior.

This whole block is worth revisiting later though for clarity.