internal/runner: unit test for accept timeout
This adds a unit test to verify accept timeouts work (they do).

At the same time, I modified our accept timeout to not cancel the entire
runner, and to avoid some data races that we had. In practice, neither
was a real issue due to the way we use accept timeouts currently. But
stepping back, I felt this was the correct behavior. Now, accept
timeouts are timeouts _per-accept_ call and does not cancel the _entire
runner_.