serverinstall/nomad: Automatically setup Consul service on install
Prior to this commit, the consul service integration was an optional
opt-in service that users could request when using the server install
helper CLI with Nomad. This commit changes that by instead setting up
the Consul service automatically, given how many Nomad workflows require
Consul for any kind of networking.