support arm64 for the entrypoint, inject properly
This adds a `linux/arm64` entrypoint asset to our Waypoint binary. This
increases binary size by ~30% but I think this is worthwhile as more of
the world's dev and deployment machines move to arm64.

This makes it so that Docker entrypoint injection injects the entrypoint
binary that matches the architecture and errors if it detects an
unsupported architecture.

This also adds detection for cloud native buildpacks to warn if the
Docker daemon isn't x86_64 since Buildpacks currently have issues with
arm64.