ui: add git commit mock data to mirage
This is cheap-and-cheerful solution, and an example of where I think
it’s worth cutting corners in Mirage. A full-fidelity solution would
include modelling and simulating `Job`, which is one of the largest and
most complex protobufs in the system, and therefore more work than is
currently justified (in my opinion). As we push further into
introspecting data about jobs, I’m pretty confident the situation the
time to dive into simulating Job will clearly present itself.

Closes #2680