serverinstall/nomad: Ensure allocations start up beyond init check
Prior to this commit, the server install command would look for the very
first instance for when an allocation was considered running. This is
generally ok, however if the Nomad job fails to get started a moment
later (like a static runner failing to connect back to Waypoint Server),
the runner allocation would exit later but the server install command
would consider the installation successful.

This commit fixes that by doing a few retries on the scheduled
allocation once its in a "running" state to ensure it properly started
up beyond the first few moments of the job.

Fixes #2683