Implement list runners
Very basic first implementation of list runners.

The immediate need for this is to determine if any remote runners exist, so we can decide if we can execute projects remotely vs locally.

This doesn't entirely meet that need - because runners aren't currently marked as remote vs cli, there's no way to tell if a runner is remote, or just some other user who happens to be running a CLI command with a local runner. For my immediate purposes though, that kind of false-positive isn't a dealbreaker. We can add more filter capabilities in the future if we need to.