Commit origins, add DOTgitignore
Each repo has it's origin as a folder within itself, .gitignored. This lets each test be self-continaed, but
the parent waypoint repo was respecting those gitignores. The parent repo needs to commit the origin folders,
but the test repos need to ignore them (otherwise they'll never be fully in sync).

The solution is to apply the same DOTgit pattern to the gitignore.