ui: rearrange deployment routes
This is a breaking change in two ways:

1. Paths that used to be like this:
   `/{workspace}/{project}/app/{app}/deployment/seq/{seq}
   Are now like this:
   `/{workspace}/{project}/app/{app}/deployments/{seq}
2. The legacy ID-based deployment routing is gone entirely

Originally I didn’t intend to make any breaking changes, but doing so
dramatically simplified the codebase so overall it seemed worthwhile.
We are pre-1.0 after all, and I think users will be able to recover
from a 404 reasonably easily. Feel free to push back on this assertion.