internal/server: Cache Horizon URL Name and Host for Deploy Bundle
Prior to this commit, when crafting a Deploy bundle, we would include
the applications deployment URL for each deployment in the bundle.
Looking up this hostname requires making an HTTP request to Horizon. The
way this was originally written, we would look up that URL for all
requested deployments. If a user had 80+ deployments in the bolt db,
this request would take quite some time.

Given that each applications deployment vanity name and hostname should
be the same, and only differs by the deployment sequence, we don't need
to get the hostname for each requested deployment. Instead we can grab
it once, and use it for the remaining deployments in the bundle.

Fixes #2304