server/httpapi: Update streamed Message format
This commit updates the crafted message format that is streamed back to
the user. Now we return a value and valueType, where value is an
interface that is the direct terminal event value, and the valueType is
the type of terminal stream event.

Update stream message example in trigger docs