ui: use alternative method to support icons in Percy
Our previous solution caused a11yAudit to take noticeable longer, due to
the presence of a duplicate spritesheet in the page for every single
test (documented in [#2934]).

This patch achieves the same ends, but does the work to include the
spritesheet during Percy’s snapshotting procedure. It’s a little more
complex, but only applies to the Percy tests and avoids adding
redundant sprite sheets to the rest of the suite.

[#2934]: https://github.com/hashicorp/waypoint/issues/2934