internal/server: Ensure on-demand runner config exists before start job
Prior to this commit, if the state package returned nil or nothing for
an on-demand runner config based on the id set for a job, Waypoint would
still attempt to queue the job with a nil ODR config. This leads to a
panic immediately after when we attempt to set environment variables
from the ODR config for the job on a nil config. This commit fixes that
by returning an error if the requested ODR config does not exist in the
state database.