cli: Fix panic in `waypoint plugin` cli
Prior to this commit, if a user just ran `waypoint plugin` with no
arguments, the CLI would panic assuming that an argument was passed in.
This commit fixes that by looking at the argument slice and printing an
error if none was passed in. It fixes up some of the other panics to
print errors and return 1.