internal/server: RunnerJobStream verifies runner tokens
This fixes a possible vulnerability with runner adoption. Note this
vulnerability is not in any released Waypoint version, so there is no
broad communication necessary.

We were not validating the runner token against the runner identity for
RunnerJobStream. So long as the token was valid, any runner could
request a job impersonating any other runner. This did require that the
runner was authenticated already.

This adds token verification and a unit test to validate it.