CLI respects server ordering
It's up to the server to choose the order (based on the -order-by and -desc flags).

NOTE: the server, in practice, doesn't respect any of the ordering flags. It only references the COMPLETE_TIME index (in app_operation), and totally ignores `-desc` and `-order-by=start-time` options.

In practice, this doesn't change any behavior - operations always come in ordered by sequence id, descending. But this opens the door for us to fix this in the server later (or for other server implementations to do this right).