internal: Don't include app if not set on job
Prior to this commit, our poll operation would set the poll jobs
singleton Id to Workspace/Project/Application. However not all project
poll operations have an application, such as the initial QueueProject
operation. In this case, we only set Workspace and Project. This means
the singletonId for the Job included a double slash, such as:

poll-trigger/go-gitops-0//default

Where the application was not set for the job.

This commit fixes this by rearranging the singleton to start at
Workspace, Project, and then optionally Application if exists on the
job.