internal: Update Task job state as runner executes it's job triple
This commit updates the Task state object to follow along with its three
jobs: Start Task, the Source Job, and Stop Task. As each job executes,
the parent Task will track each of its three job states and represent it
as its own Task State. As a runner gets to each job in the task, it will
make a call to the server to update this state during the life of a
Task.