ci: Only notify PR check failures if `website/content` has changes
Prior to this commit, any changes at all from running `make
gen/website-mdx` would notify a failure. This is extra tricky because
npm and lock files might change which causes `git diff` to notify the PR
that a failure occured. In reality, this check was only to tell the PR
author that they needed to run `make gen/website-mdx` to generate CLI
docs automatically. This commit updates that behavior to only check a
`git diff` in the `website/content` folder, where the auto generated
docs live.