pkg/server: fix flaky test around buffered data
To fix this, we need to retry a connection until we get buffered data.
We have to do this because the buffered data is read on connection and
immediately queued to send. It is not dependent on the client receiving
(though, it'll block due to TCP backpressure if they aren't receiving).