cli: Show error message if no Waypoint contexts found
Prior to this commit, if Waypoint had no contexts locally on disk, it
would still attempt to load one with an empty name, resulting in an
error about a nil context. This commit fixes that by first seeing if any
contexts exist at all, or if there's no default, showing a better error
message to tell the user what to do about it.

Fixes #3243