runner: static runners accept multiple jobs in parallel
This modifies `internal/runner` to support accepting multiple jobs in
parallel. Not much work here since we always designed the runner struct
from the beginning to support this so there are no data races.

This modifies `internal/cli` so that runners in non-ODR run in parallel
mode by default. ODR doesn't make sense to have any parallelism since
they always run exactly one job. Non-ODR runners typically ONLY launch
ODR tasks, which are highly IO-bound, so we default to a multiple above
CPU count for concurrency.

This is a necessary pre-requisite for pipelines since they'll likely
perform blocking jobs on the static runners to "watch" tasks. Today,
tasks are launched and stopped, but not watched so this is not an issue.