builtin/k8s/task: Explicitly set container env var for Kaniko
Prior to this commit, if a certain version of Kaniko was used with the
current latest version of Kubernetes 1.23, Kaniko would fail to detect
that it was being run inside a container. This means Waypoint would then
faill back to try to use Docker inside a container for the build step.
This commit fixes this by setting the proper env var for Kaniko and
keeps ODR remote builds working.

Fixes #2984