server: Ensure primary ODR job gets added as dependency
Prior to this commit, there was a bug with ODR and generating the Stop
Task job. If a plugin did not implement WatchTask, the WatchTask job would fail,
which would invoke the StopTask job and break the task state machine while the
main job was still running. This commit fixes that behavior by ensuring
that if a source job has no overrides defined and is not expected to
skip the operation, the source id gets added as a dependent for the stop
task.