core/runner: Add runner k8s uninstaller
This commit implements the runner uninstaller for kubernetes using helm.
When uninstalling a runner with waypoint the platform and runner ID are
required arguments.

This commit also takes care of cleaning up persistent volumes left
behind by the initial helm installation so that runner uninstallations
leave the environment in the same state that they were in previous to a
`waypoint runner install` command.

Before running an uninstall for runner we check to make sure that the
helm chart we are uninstalling matches something we would expect to be a
runner install as a sanity check.