core/runner: Add keep-pv flag to runner uninstallations
Before this commit we assumed that users would always want to clean up
their persistent volumes during an uninstallation. As this is not
cleaned up by helm be default we should at least allow the user to
specify that they would like to keep that behavior in the event they
would like that volume to not be cleaned up.