internal/server: Include `WatchTask` in task state
This commit includes the WatchTask job for the task state machine that
tracks the lifecycle of spawned on-demand runner tasks in a given
Waypoint job. The task state machine needs no changes, as `WatchTask`
simply watches the run task for log streaming. Given that, this commit
simply includes the watch task so it can be inspected via the CLI and
other operations that look at ODR tasks on a given job.