kaniko: Include registry host when returning error in entrypoint
Prior to this commit, when a user was confronted with an error about
setting up an entrypoint, it didn't give much context as to what went
wrong. This was often because the user was pushing to a remote registry
which was unexpected. This commit updates that behavior to include the
registry host in the returned error message