internal/server: Ensure k8s task plugin properly sets install profile
Prior to this commit, when the server install function went to json
marshall its config into a k8s task config, it would translate the k8s
config to be `Requested`, but the k8s platform config expected `request`
and would fail to parse the profile. This commit ensures that the task
runner profile properly sets request for both mem and CPU so that it can
be parsed.