More error wrapping around ops
Due to an unrelated bug in the waypoint server, we were exiting early when creating status reports. The only error the client cound see was:

```
unknown error
```

Which is not helpful. This wrapps the errors in that flow to give more context as to what the client was attempting to do when it encountered the error.