builtin/docker: Do not require `resources` for task launch config
Prior to this commit, the task launcher was expecting `Resources` to be
a required plugin config option despite it not actually being required.
This was due to the `TaskResources` field being expected to exist in a
task plugin despite being optional. The fix, similar to how we handle
resource plugin configs in kubernetes, is to make the Resources block a
pointer so that when it's unset, the plugin config parser doesn't expect
it to be required.

Fixes #3485