internal/{server,runner}: Queue a Release job via a pipeline step
This commit updates the pipeline run service API func to queue a Release
job if requested from a pipeline step. Additionally, because pipelines
is expecting to support referencing "latest" deployments as well as
deployments by sequence numbers, the runner impl for running a Release
will now look up the latest deployment if the Job operation
configuration for a Deployment was not set i.e. nil.