internal/runner: Use `target` rather than the Deployment on the job operation
Prior to this commit, the operation release for a runner would operate
assuming that a jobs Release operation would always have a Deployment
proto attached to it. This changes in the pipelines world, because we
allow for Deployment Refs for queuing jobs ahead of time. Prior to even
pipelines, the runner release operation already was extracting out the
deployment target from the Release operation, but we'd still try to use
the deployment on the job proto instead.

This commit resolves that by making sure the release runner code uses
target.