maint: Add better error message when runner unable to connect
Before this commit when a runner was unable to connect to the server to
be adopted the message was vague and didn't indicate a specific issue
with the runner. This commit should hopefully add additional context as
to why the runner may not have been able to connect to the server.