k8s/task: Return an actual error if pod job exists but has no pods
Prior to this commit, if Waypoint server happens to panic after it
attempts to launch a task, but the task was unsuccessful and the pod job
has no actual pods, the WatchTask would panic because it assumed if a
pod job exists it would have pods inside. This commit adds an extra
check for the length of the pod items prior to attempting to grab a pod
to watch.