Fix ordering of service names for ECS runner uninstall.
The AWS SDK can possibly return a service for a cluster that no longer exists. Because of this, the FindServices call in Uninstall for the ECS runner finds the 'ghost' service and attempts to uninstall that, never deleting the actual intended service. Because of this, the ordering of the service names has been fixed to check for the "old" runner service name first, waypoint-runner.