server: Build dependency graph across pipelines prior to Run
This commit builds the full Pipeline run graph with associated jobs
prior to running the pipeline. It will also detect cycles between
embedded pipelines if any exist. To prevent cycles while building a
dependency graph, we include visisted nodes and pipelines so that if we
visit a node we've already been to, we simply move on rather than
continuing through the cycle