server: Handle building graphs with same node names
Prior to this commit, building a graph of pipelines and embedded
pipelines had an issue if the nested pipeline steps had step names that
matched any other step from another pipeline. This in practice is valid
to have step names be similar to other pipelines but when we combine
them and use them to constrct a graph to determine dependencies, we
can't re-use that name because we expect node names in the graph to be
unique.

This commit fixes this by assigning node IDs to each node in a graph,
and storing a pipeline step ref to node Id map that we can look up later
when we need the step and pipeline names.