Set destroyed and remaining resources after a destroy op runs.
The proto message used in the execution of a deployDestroyOperation now sets two fields of a Deployment proto message, DeclaredResources and DestroyedResources. These are available to the caller that initiated the operation with doOperation. The data in these fields is used to output the resources that were and were not destroyed during the destroy operation.