Move passing waypoint-token not in Authorization header
The majority of this change is refactoring and DRYing up the token
handling.

But the critical bit is that it begins the process of moving the
waypoint token away from the Authorization header and to a new
Waypoint-Token header in GRPC. The reason for doing this is that,
technically, we're using the Authorization header wrong anyway, we don't
put a type before the value, which is the proper convention.

Additionally, by moving the token to it's own header, the system can
perform a different primary authentication (such as with oauth) via the
Authentication header and still pass the token along side, which can be
used for authorization.