ignore unavailable errors
That's normal apparently in our integration tests, and this is causing one of those tests to fail. Won't argue with it.

We don’t _need_ to warn log anything there. If we hit that condition, we’ll never send the inline keepalive messages. The log is nice in case someone starts reporting the the RST_STREAM errors again - we can ask for this log and figure out what the problem was. This isn’t worth more than one more round of whack a mole though. If we find another case like this, we should switch this to debug/trace log.