internal/core: Ensure proto message is proper type before return
Prior to this commit, if any of the following plugins
build,deploy,artifactpush, or release did not properly return the
expected protobuf message, the internal core funcs would attempt to
convert the proto mesage without ensuring it was the proper expected
type. If this ever happened, the core func would panic and break
completely. This commit fixes that with some extra guarding to first
check if we can properly convert the message to our expected proto type,
and if not, return an internal error code rather than panicing.