Fix getting which app operations to delete during project delete.
Prior to this commit, the slices of operations to be deleted were being overwritten with each app. This commit updates the behavior to append them to a slice to be deleted later, so the operations for all apps in a project are retrieved and then deleted. A test is added for this behavior as well.