Delete project-associated records in one transaction.
This commit enables app operations to be deleted with a new function 'delete' that accepts a DB transaction as an input parameter. This allows the delete logic for each of them to be called during an existing transaction without causing a deadlock.

Furthermore, non-app operation records, like pipelines or triggers, are now similarly called, in that the delete function for each, which allows a transaction to be passed as an input parameter, is used.