App operation updates.
This commit updates the deletion of app operations to also be deleted from the in-memory database. Furthermore, when an operation is deleted, its sequence # is decremented. This is necessary because if a project is re-initialized after deletion, new operations would have a sequence # set to the sequence # prior to the project's deletion, plus one.

In order to do this, the app and project name for each operation is required. Thus, the full proto is passed to the delete function for each app operation, instead of just a Ref_Operation.