Making keepalive send errors quieter.
With the old log level, if there's a problem with the server connection, we'd usually log _two_ errors - one for the currently active stream (like RunnerConfig), and also an inline keepalive error.

The keepalive error is really just noise for the user in that case, and the client failing to send a keepalive isn't something a user should generally act on, so i'm downgrading this to debug level.

I've also noticed this log-erroring sometimes on shutdown. I think it's possible for us to call SendMsg in the keepalive routine after CloseSend has been invoked elsewhere, but before the context is cancelled, which yields us a grpc Internal type error from SendMsg. We could avoid this with more mutexes, but I don't think it's worth the complexity to avoid.