Update ODR config interface for server and runner installers.
This commit updates the OnDemandRunnerConfig interface to accept no arguments once again. The runner install CLI appends the ID of the runner to the profile name after this config is retrieved. Also, if target labels for the runner profile are specified (not possible yet), then the profile uses target labels instead of a target ID.