core/runner: Fix server panic when runner is forgotten
Before this commit when Waypoint was determining whether or not to
remove a Runner from boltdb it was possible for runner to be nil at the
time we attempted to determine what type of Runner the Runner was. This
caused the server to panic as soon as the runner became unavailable.

This commit fixes the panic by checking if we received a runner from the
database before determining its type.

Fixes #3448