Cycle through workspaces during project destroy operation.
This commit adds a new function to the project type in core to create a copy of a project with a different workspace. The project destroy operation requires this because when the workspace is set by the runner, executing a destroy operation for an app is only possible in that workspace, regardless of the workspace proto that is set on the job for a destroy operation.

The Waypoint config is required to be plumbed through because the config is needed to re-init a project when we get a copy of it, so we provide it to the project destroy operation from the runner.