Delete runner profiles by name
Closes https://github.com/hashicorp/waypoint/issues/3481

This allows runner profiles to be deleted by name, not ID. We don't surface IDs to users, so it's not useful to delete them by ID.

I also found that we were able to create multiple runner profiles with the same name. This PR fixes that, adds a test around it, and still allows users to delete runner profiles by `id` (via a flag) if they find themselves with duplicate-named runners as an emergency backdoor.