cli: Warn if local invocation of `waypoint pipeline run`
Prior to this commit, the CLI would attempt to run a pipeline locally if
the `-local` flag was used. Currently the MVP of pipelines does not
support this, so we should show a big warning as such prior to running
the pipeline so that users aren't confused when they see an error about
the project not being setup to run remotely.

Fixes #3811