Update version check on server during ECS upgrade.
Previously during ECS upgrades, if the current server image was hashicorp/waypoint:latest, regardless of user input, the upgrade would just force-redeploy the existing ECS service and not update the image. With this change, we check the existing image to see if it matches hashicorp/waypoint:latest, AND check if that's what the user wanted to upgrade to. If either of those is false, we update the task definition during the upgrade.