server: Track embedded step refs for pipelines with noop job
Prior to this commit we would ignore step refs of type pipeline and not
generate its own job. This caused an issue where if a embedded pipeline
steps parent step was also an embedded pipeline, we wouldn't have a job
id to wait on, which would make the step run out of order from its
parent.

This commit fixes that by tracking the step refs as a noop job, then
that job depends on all of its embedded jobs before it's complete. This
makes it easier for any consuming steps to wait on the noop job which is
essentially the last job of a pipeline.