Fix the jobspec plugin status check for canary deployments.
The Noamd jobspec platform plugin status check would incorrectly report a deployment as partially healthy, if the Nomad job had canary allocations. It is possible for such a deployment to be partially healthy; however canary allocations end up in the "complete" state when the deployment is promoted. Therefore, after using the "release" lifecycle with the canary plugin, the # of "complete" allocs would be > 0, and thusly cause a partially healthy state, regardless of the fact that the app is indeed healthy, and its deployment was just promoted, killing the canaries.