feat(plugins/docker): add docker-ref plugin
The `docker-pull` plugin really pulls the image in the case of On Demand Runners.
Since we do require there a `docker.AccessInfo`, we can't avoid pulling the image
(because HasRegistry automatically becomes true, and not providing a registry causes an error).

To avoid patching `docker-pull` further, a `docker-ref` is created.
This is only used in the very specific case where a `docker-pull` can't be used because all of the following
conditions are met:

- Access to the registry is not available
- On-Demand Runner (ODR) is being used

As a solution, this plugin allows you to be a sort of "noop" that adds a reference to a pre-built
Docker image (e.g: in an external CI) in the `build` stage.