cli: Fix panic when no flag specified for `waypoint runner profile set`
Prior to this commit, if you ran the profile set command without any
flags, the command would panic due to the flagPluginType being a string
pointer. This commit fixes that by instead moving it back to a simple
String flag with a emptry string default. It preserves the behavior
where if not specified, the CLI will attempt to use the existing plugin
type rather than setting it to empty string.

Fixes #4012