pipeline: Properly determine when pipeline run is complete
Prior to this commit, when we'd go to complete a pipeline run, we would
only look at the "last" job in the jobs slice for a pipeline run
message. This check doesn't seem to work anymore, and relying on the
last job in a slice can be fragile. This commit fixes this behavior by
checking all job ids in a pipeline run and seeing if they have finished
as well before attempting to mark the pipeline run as a success.

Fixes #4032