internal/runner: Run a status report on Deployments and Releases
Prior to this commit, a status report operation was not auto-executed
any time a deployment or release ran. This meant as Waypoint developers
we had to be explicit when we wanted Waypoint to generate status
reports. Given that our CLI runner auto-executes status reports on up,
deploy, and release, it makes sense other systems that use these
operaitons should also get status reports in an easy way. This commit
updates the behavior of deploy and release to run a status report at the
end of its execution so that we can get some simple checks after
deploying or releasing.