cli: Only echo conf file on fmt change
Prior to this commit, the `waypoint fmt` command would echo the
formatted file name even if the file was not formatted. This commit
adjusts that to align the CLI to behave like `terraform fmt` where it
only echos the file name in the case where the file was actually
formatted.

Fixes #3816