internal/config: Return err if no scoped use stanza found
Prior to this commit, if you attempted to use a build, deploy, or
release stanza that had a workspace scoped 'use' stanza but you were not
using that workspace, Waypoint would panic attempting to evaluate the
Use plugin type. This commit fixes that by first looking to see if a Use
stanza is defined at all prior to obtaining the Use type.

Fixes #3759