cli: Only show CompleteTime if set
Prior to this commit, the `waypoint pipeline list` CLI would attempt to
show an invalid complete time, even if not set on the job. This would
mean if you started a pipeline, the list CLI would claim that the
pipeline completed "a long time ago", which is really code for that
timestamp being nil. This commit fixes that behavior by only setting the
complete time string if set on the job.

Fixes #3748