cli/pipeline_run: Include variables in job template for pipeline run
Prior to this commit, when a pipeline was run, the CLI variables from
the job client were not included in the submitted job template that the
CLI would send to the server to run pipeline jobs. This means any CLI
variables would not be properly resolved when the job went to execute.
This commit fixes that by including those variables in the job template.

Fixes #4134