fix: if doing a server upgrade, update the OCI URL.
Prior to this commit, the server upgrade command would use the installRunner function to install a new runner during the upgrade with the new version, AND update the runner profile to use the config defined in server_upgrade.go. However, the update of the runner profile never happened, because existingODRConfigSetup was set to true, skipping the code which would update the profile.