fix: update config source record unique ID.
The unique ID of a record in the config source "table" is no longer the "type" of config source. Instead, it is now the hash of the config source's type, scope, and configuration. This makes each entry that is added to the table unique, and thusly not overwritten by other config sources of the same type or scope.

This commit also adds another test, and fixes the ordering of params for require.Equal throughout the config source tests.