fix: Update config source record ID to be the hash of the scope and type.
The previous commit made the unique ID of a config source the hash of a config source's type, scope and configuration. This commit removes the configuration from that hash, because a change to a config source's configuration should not result in the creation of a new record, but the updation of an existing one for a specific "type" of plugin in the same "scope".

To accommodate this, the config source project index schema was also updated to include an indicator that a config source was specifically for the project level. This is because app-scoped config sources also have the project set, so the project index was matching on project-scoped AND app-scoped config sources - it now will only match on project-scoped config sources.